Change junit

+++ src/test/java/interview/lists/SinglyLinkedListTest.java
@@ -16,21 +16,21 @@ public class SinglyLinkedListTest {
     assertNull(list.head);
     list.addFirst(4);
     assertNotNull(list.head);
-    assertEquals("UnexpectedDifferentValue", 4, list.head.getValue().longValue());
+    assertEquals("UnexpectedDifferentValue", 4, list.head.getValue().intValue());
 
     list.addFirst(3);
     assertNotNull(list.head);
-    assertEquals("UnexpectedDifferentValue", 3, list.head.getValue().longValue());
+    assertEquals("UnexpectedDifferentValue", 3, list.head.getValue().intValue());
     assertNotNull(list.head.getNext());
-    assertEquals("UnexpectedDifferentValue", 4, list.head.getNext().getValue().longValue());
+    assertEquals("UnexpectedDifferentValue", 4, list.head.getNext().getValue().intValue());
 
     list.addFirst(2);
     assertNotNull(list.head);
-    assertEquals("UnexpectedDifferentValue", 2, list.head.getValue().longValue());
+    assertEquals("UnexpectedDifferentValue", 2, list.head.getValue().intValue());
     assertNotNull(list.head.getNext());
-    assertEquals("UnexpectedDifferentValue", 3, list.head.getNext().getValue().longValue());
+    assertEquals("UnexpectedDifferentValue", 3, list.head.getNext().getValue().intValue());
     assertNotNull(list.head.getNext().getNext());
-    assertEquals("UnexpectedDifferentValue", 4, list.head.getNext().getNext().getValue().longValue());
+    assertEquals("UnexpectedDifferentValue", 4, list.head.getNext().getNext().getValue().intValue());
 
 
     System.out.println(list);
